Add files via upload
This dataset is used in conjunction with the Global COVID-19 Policies dataset. Each country has its own data file. This can be used to see the data from which the dashboard was created and it can be used a template to recreated dashboard for other countries. Included in this dataset are date, daily cases, daily deaths, cumulative cases, cumulative deaths, policy name, transmission type, policy category, week, and policy ranking. The pie chart was created by manually determining frequency of transmission type.